test(platform-wallet): move ovk_recovery_tests to its own file
Extracts the inline `#[cfg(test)] mod ovk_recovery_tests` out of sync.rs into
sync/ovk_recovery_tests.rs (a file submodule). The doc comment stays on the
`mod ovk_recovery_tests;` declaration; `super::` references resolve to the same
`sync` parent as before. No test logic changed.
